refactor: move functions around
